Undo changes to `GrapheneCursor` and priority type
The cursor seems unused, the priority is checked in code elsewhere.

+++ python_modules/dagster-graphql/dagster_graphql/schema/instance.py
@@ -114,7 +114,7 @@ class GraphenePendingConcurrencyStep(graphene.ObjectType):
     stepKey = graphene.NonNull(graphene.String)
     enqueuedTimestamp = graphene.NonNull(graphene.Float)
     assignedTimestamp = graphene.Float()
-    priority = graphene.BigInt()
+    priority = graphene.Int()
 
     class Meta:
         name = "PendingConcurrencyStep"
