+++ python_modules/dagster-graphql/dagster_graphql/schema/paging.py
@@ -1,6 +1,6 @@
 import graphene
 
 
-class GrapheneCursor(graphene.BigInt, graphene.Scalar):
+class GrapheneCursor(graphene.Int, graphene.Scalar):
     class Meta:
         name = "Cursor"
